Add routes and searchStock
