+++ src/components/Nav.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import nav from '../style/nav.module.css';
+import SearchStock from './SearchStock';
 
 const Nav = () => (
   <nav className={nav.nav}>
     <div><a href="/">RecipesEase</a></div>
     <div className={nav.right}>
       <a href="/category">Category</a>
-      <a href="/meals">Meals</a>
       <a href="/About">About</a>
+      <SearchStock />
     </div>
   </nav>
 );
