+++ src/components/RecipeCategory.jsx
@@ -2,17 +2,25 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import NumberFormat from 'react-number-format';
 import { Link } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
 import allstocks from '../style/allstocks.module.css';
+import { getRecipe } from '../actions';
 
 const RecipeCategory = ({ stock }) => {
+  const dispatch = useDispatch();
+
   const {
     symbol, companyName, price, volume, sector,
   } = stock;
+
+  const handleClick = () => {
+    dispatch(getRecipe(symbol));
+  };
   return (
     <>
 
       <div className={allstocks.box} key={symbol}>
-        <Link to={symbol}>
+        <Link to={`/recipe-details/${symbol}`} onClick={handleClick}>
           <div>
             <h2 className={allstocks.title}>
               {' '}
